Rename interactors & functions
